// src/types/swagger.ts
export interface ExternalDocs {
  url: string;
  description?: string;
}

export interface Tag {
  name: string;
  description?: string;
  externalDocs?: ExternalDocs;
}

export interface Header extends BaseSchema {
  type: string;
}

export interface Ref {
  $ref: string;
}

// ----------------------------- Parameter -----------------------------------
export interface BaseParameter {
  name: string;
  in: string;
  required?: boolean;
  description?: string;
}

export interface BodyParameter extends BaseParameter {
  schema?: Schema;
}
// ...
export interface BaseSchema {
  format?: string;
  title?: string;
  description?: string;
  default?: string|boolean|number|{};
  multipleOf?: number;
  maximum?: number;
  exclusiveMaximum?: number;
  minimum?: number;
  exclusiveMinimum?: number;
  maxLength?: number;
  minLength?: number;
  pattern?: string;
  maxItems?: number;
  minItems?: number;
  uniqueItems?: boolean;
  maxProperties?: number;
  minProperties?: number;
  enum?: Array<string | boolean | number | {}>;
  type?: string;
  items?: Schema|Schema[];
}

export interface Schema extends BaseSchema {
  $ref?: string;
  allOf?: Schema[];
  additionalProperties?: Schema;
  properties?: {[propertyName: string]: Schema};
  discriminator?: string;
  readOnly?: boolean;
  xml?: XML;
  externalDocs?: ExternalDocs;
  example?: any;
  required?: string[];
}

export interface XML {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
}